Add some more structure and an additional test to get things going
